refactor conversation creation
